infrastructure/cdk.out/asset.bbf47a15fe72a4b2c9f2ecd03c890e45b6e5ba0ddd5ed9914248092bbbd415cf/spacex_site/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework import permissions
from drf_yasg.views import get_schema_view
from drf_yasg import openapi
from django.http import JsonResponse

schema_view = get_schema_view(
    openapi.Info(
        title="SpaceX Launches API",
        default_version='v1',
        description="API para acceder a datos de lanzamientos de SpaceX contenidos en DynamoDB",
        contact=openapi.Contact(email="martinezjorge960616@example.com"),
        license=openapi.License(name="MIT License"),
    ),
    public=True,
    permission_classes=(permissions.AllowAny,),
)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include('launches.urls')),
    # Swagger URLs
    path('swagger/', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
    path('redoc/', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
    path('swagger.json', schema_view.without_ui(cache_timeout=0), name='schema-json'),
    #Heald  check 
    path('health/', lambda request :JsonResponse({'status':'healthy'})),
]

