infrastructure/cdk.out/asset.bbf47a15fe72a4b2c9f2ecd03c890e45b6e5ba0ddd5ed9914248092bbbd415cf/requirements.txt
asgiref==3.10.0
attrs==25.4.0
aws-cdk-lib==2.131.0
aws-cdk.asset-awscli-v1==2.2.257
aws-cdk.asset-kubectl-v20==2.1.4
aws-cdk.asset-node-proxy-agent-v6==2.1.0
boto3==1.34.0
botocore==1.34.162
cattrs==25.3.0
constructs==10.3.0
Django==5.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
importlib_resources==6.5.2
inflection==0.5.1
jmespath==1.0.1
jsii==1.119.0
packaging==25.0
publication==0.0.3
python-dateutil==2.8.2
pytz==2025.2
PyYAML==6.0.3
s3transfer==0.9.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typeguard==2.13.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.6.0